added file for Context, optimized code
